refactor(llmclient): extract circuit breaker into peer file
internal/llmclient/client.go was 848 lines; the circuitBreaker struct,
its state constants, and its methods were a self-contained block at
the tail. Move them to circuit_breaker.go (same package, no API
change). Pure relocation: client.go drops to 720 lines.
